[S09P31A202-200] fix: PloggingLog coin 필드 타입 Integer로 변경

+++ backend/zupzup/src/main/java/com/twoez/zupzup/plogginglog/domain/PloggingLog.java
@@ -43,7 +43,7 @@ public class PloggingLog extends BaseTime {
     private Integer gatheredTrash;
 
     @Column(nullable = false)
-    private Long coin;
+    private Integer coin;
 
     @Column(nullable = false)
     private String routeImageUrl;
@@ -63,7 +63,7 @@ public class PloggingLog extends BaseTime {
             LocalDateTime endDateTime,
             Integer calories,
             Integer gatheredTrash,
-            Long coin,
+            Integer coin,
             String routeImageUrl,
             Boolean isDeleted,
             Member member) {
